pyproject.toml
[project]
name = "deepseek-api"
version = "0.1.0"
description = "DeepSeek API client using OpenAI SDK"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
